Update no birth-cert outcome for Iran
The outcome for Register a Birth where the child does not have a birth
certificate and the parents are not married is changed for Iran and
should match the outcome for same circumstances as for Iraq.

The COUNTRIES_WITH_BIRTH_REGISTRATION_EXCEPTION constant is updated to
enable this. This constant also affects marriage-abroad and
register-a-death Smart Answers, hence why in this commit the checksums
are also updated for those two.