Split out un-publishing of Smart Answer task
For the retirement of the Overseas Passport Smart Answer we need to do
this in stages. To help this we are splitting out the series of steps
in the existing rake task into separate rake tasks.

This is the first change where the un-publishing step now has it’s own
rake task.