Reserve path for publishing application
The commit adds a method that reserves the given base path for a
publishing application.

This is based on the publishing api call [1] and in this context is it
used  to switch the publishing application from a smart answer to another
publishing application.

The put_json and reserve_path_url have been used because there is no
method implementation for this call [1] defined in the gds-api-adapters
gem [2]

[1] https://github.com/alphagov/publishing-api/blob/master/doc/api.md#put-pathsbase_path
[2] https://github.com/alphagov/gds-api-adapters/blob/88df252df87526c0ee269fb2586d926e0fc86e7c/lib/gds_api/publishing_api_v2.rb