Add rake task for changing publishing app
Add a rake task that calls the ContentItemPublisher method that reserves
the publisher application path for a Smart Answer. This is so that we
can re-assign a Smart Answer that is being retired, but may have a
start page owned by another app at the base-path.

This is to allow us to retire the Overseas Passports smart answer which
will move to a new passports service, but the start page will exist on
the Mainstream Publisher app.